refactor(#618): [Vijay|Alka] refactor storage method and variable name to follow convention
