+++ shared/storage.ts
@@ -1,5 +1,5 @@
 import { MMKVLoader } from 'react-native-mmkv-storage';
-import { VC_ITEM_STORE_KEY_REGEX } from './shared/constants';
+import { VC_ITEM_STORE_KEY_REGEX } from './constants';
 import {
   DocumentDirectoryPath,
   readDir,
@@ -9,9 +9,9 @@ import {
 } from 'react-native-fs';
 
 const MMKV = new MMKVLoader().initialize();
-let regExp = new RegExp(VC_ITEM_STORE_KEY_REGEX);
+let vcKeyRegExp = new RegExp(VC_ITEM_STORE_KEY_REGEX);
 export const getDataFromStorage = async (key: string) => {
-  if (regExp.exec(key)) {
+  if (vcKeyRegExp.exec(key)) {
     const path = getFilePath(key);
     return await readFile(path, 'utf8');
   }
@@ -19,7 +19,7 @@ export const getDataFromStorage = async (key: string) => {
 };
 
 export const setDataToStorage = async (key: string, data: string) => {
-  if (regExp.exec(key)) {
+  if (vcKeyRegExp.exec(key)) {
     const path = getFilePath(key);
     return await writeFile(path, data, 'utf8');
   }
@@ -32,11 +32,11 @@ export const clearDataFromStorage = async () => {
   MMKV.clearStore();
 };
 
-const getFileNameFrom = (key) => {
+const getFileName = (key: string) => {
   return key.split(':').join('_');
 };
 
 const getFilePath = (key: string) => {
-  let fileName = getFileNameFrom(key);
+  let fileName = getFileName(key);
   return `${DocumentDirectoryPath}/${fileName}.txt`;
 };
